requirements.txt
langchain-google-genai
langchain-chroma
langchain-community
langchain-experimental
langgraph
pypdf
chroma
dotenv
